rag_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
faiss-cpu==1.8.0
sentence-transformers==3.1.0
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.0
tf-keras
